src/flash/nor/mspm0.c: Update mspm0_protect to have write/read checks
Target_write/read_u32() was never read which could lead to unknown errors
to lurk. With this commit a check is performed upon any read/write to
the device within the protect() function. If the read fails it is
allowed to continue otherwise if a write fails then there must be
a severe error and it is not allowed to continue.

Change-Id: I62c9d721c62f1885e49dff307da848b7d5c2a753
Signed-off-by: Henry Nguyen <h-nguyen8@ti.com>