doc/openocd.texi: Update documentation to include expected behavior post-reset while empty
User's could potentially go in and reset the device post-erase while the MCU is still
executing from MAIN memory. Which can cause issue as the interrupt vector table is
missing if all of MAIN is erased leading to a double hard-fault. If someone is unaware
and did so it could lead to bug reports for a known behavior.

Change-Id: Iaeb4e69e804538ebe57354d049fabc35601035cd
Signed-off-by: Henry Nguyen <h-nguyen8@ti.com>